Fix documentation links for JS module features
The first is "JavaScript module import()" which is also referred to as
"dynamic imports". It differs from ES6 modules in that it allows
importing code at runtime (whereas ES6 modules only enable static
linking). I've updated the description to clarify this. The previously
used MDN link only discussed static imports. I've updated it to point
to the proposal on GitHub (as already present in the "link" field)
since no suitable MDN page exists for this yet.

The second is "Modules (ES6)" which is the static import syntax. This
was pointing to an unrelated "JavaScript code modules" documentation
page on MDN which has nothing to do with this standard. I've updated
it to point to the MDN documentation for import instead.